# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления бронированиями",
        "title": "Booker Microservice API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/bookings/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список всех объектов бронирований с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Получить список бронирований у пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Сортировка по id. asc, desc",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка по booking_entity_id. asc, desc",
                        "name": "booking_entity_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка по start_time. asc, desc",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка по end_time. asc, desc",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка по status. asc, desc",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка по user_id. asc, desc",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Лимит на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking.BookingsList"
                        }
                    }
                }
            }
        },
...
        "/bookings/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить детальную информацию о бронировании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Получить бронирование по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking.BookingInfo"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновить бронирование",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Обновить бронирование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные бронирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_create_booking.BookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.ResponseId"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить бронирование по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Удалить бронирование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/bookingsEntity/{id}/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список всех бронирований с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Получить список бронирований у объекта бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объекта бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Лимит на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking.BookingsList"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ShlykovPavel_booker_microservice_models_booking.BookingInfo": {
            "type": "object",
            "properties": {
                "booking_entity": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking.BookingsList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking.BookingInfo"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking.BookingsListMetaData"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking.BookingsListMetaData": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking_create_booking.BookingRequest": {
            "type": "object",
            "properties": {
                "booking_entity_id": {
                    "type": "integer"
                },
                "companyId": {
                    "type": "integer",
                    "format": "int64"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.ResponseId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Add \"Bearer\" before token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}